hostapd: Remove recipe and replace with bbappend
OE just updated hostapd to version 2.2, so the full recipe is no longer needed. We
still need to modify the installation in order to keep the modified hostapd.conf
and to enable hostapd by default, so we will keep these bits.  Other redundant files
will be deleted in a subsequent commit.

Also, the recipe was renamed from hostap-daemon to hostapd in meta-oe, so rename
the addtional files directory to a generic name in order to make it easy to work
with subsequent version bumps of hostapd.